[LLDB][ELF] Load both, .symtab and .dynsym sections
Summary:
This change ensures that the `.dynsym` section will be parsed even when there's already is a `.symtab`.

It is motivated because of minidebuginfo (https://sourceware.org/gdb/current/onlinedocs/gdb/MiniDebugInfo.html#MiniDebugInfo).

There it says:

> Keep all the function symbols not already in the dynamic symbol table.

That means the `.symtab` embedded inside the `.gnu_debugdata` does NOT contain the symbols from .dynsym. But in order to put a breakpoint on all symbols we need to load both.

To not add symbols that where already added I keep a list of unique ELF symbols on each invocation of `ObjectFileELF::GetSymtab`.

My other patch D66791 implements support for minidebuginfo, that's why I need this change.

Reviewers: labath, espindola, alexshap

Subscribers: JDevlieghere, emaste, arichardson, MaskRay, lldb-commits

Tags: #lldb

Differential Revision: https://reviews.llvm.org/D67390

llvm-svn: 371599